+++ R/getProjectsCM.R
@@ -22,8 +22,8 @@
 #' @details 
 #' Additional arguments: 
 #' 
-#' \code{server}: a character to select from which server the data will be retrieved, either 
-#'  "prodution" (the default) or "testing"
+#' \code{server}: a character to select from which server the data will be retrieved,
+#'   "production" (the default) or an alternative server 
 #'  
 #' @examples
 #' \dontrun{ 
@@ -48,15 +48,11 @@ getProjectsCM <- function(key = NULL, ...){
     server <- "production"
   }
   
-  if (server == "production") {
+  url <- "https://climmob.net/climmob3/api/readDataOfProject?Body={}&Apikey={}"
+  
+  if (server != "production") {
     
-    url <- "https://climmob.net/climmob3/api/readProjects?Apikey="
-    
-  }
-
-  if (server == "testing") {
-    
-    url <- "https://testing.climmob.net/climmob3/api/readProjects?Apikey="
+    gsub("https://", paste0("https://", server, "."), url)
     
   }