+++ NEWS.md
@@ -1,12 +1,18 @@
-ClimMobTools 0.3.6.9001 (2020-12-08)
+ClimMobTools 0.3.7 (Unreleased)
 =========================
 
 ### IMPROVEMENTS
+
 * Method for `as.data.frame()` now handles ClimMob data without the assessment info.
+* `as.data.frame()` decodes values from categorical variables
+* `as.data.frame()` decodes ties and not observed values in the ranking data 
+* Handle alternative server addresses
 
 ClimMobTools 0.3.5 (2020-05-08)
 =========================
+
 ### CHANGES IN BEHAVIOUR
+
 * Imports 'PlackettLuce' and 'climatrends' by default
 * Use `httr::RETRY()` instead of `httr::GET()` as suggested by Anna Vasylytsya
 * A `print()` method is added